fix(LottoResult): 1등이 2번 당첨될 경우를 대비해 long 타입으로 변경

+++ src/main/java/lotto/domain/LottoResult.java
@@ -25,13 +25,13 @@ public class LottoResult {
 
     public double calculateRate(final Payment payment) {
         int money = payment.getPayment();
-        int totalMoney = getTotalMoney();
+        long totalMoney = getTotalMoney();
         return (double) totalMoney / money;
     }
 
-    private int getTotalMoney() {
+    private long getTotalMoney() {
         return Arrays.stream(Rank.values())
-                .mapToInt(rank -> rank.getMoney() * lottoResult.get(rank))
+                .mapToLong(rank -> (long) rank.getMoney() * lottoResult.get(rank))
                 .sum();
     }
 }
